refactor(physical): 이동 버퍼 단일 슬롯화 + 샌드박스 연속이동·착수 개선
이동 입력 버퍼를 FIFO 큐 → 단일 슬롯(최신 입력 우선)으로 정리하고,
샌드박스(솔로) 전용 연속(부드러운) 이동을 추가한다.

- 버퍼 리팩터: PhysicalPlayer moveQueue(Deque) 제거 → bufferedMove 단일 슬롯,
  moveQueueMax 설정 삭제(미리 여러 번 눌러도 옛 방향이 잔여 이동으로 튀지 않음)
- 연속 이동: 실수 좌표(fx,fy)+moveContinuous, 엔진 advanceContinuous(속도×틱, 칸 충돌),
  스냅샷 좌표 double화. 솔로 방만 활성(실대전/AI연습은 격자 유지)
- 연속 속도 배율 1.15, place-cooldown 150→110(연타 반응성)
- 착수 버퍼: 쿨다운 중 무조건 버퍼 + 착수 불가 칸이면 window 내 유지(연타 씹힘 방지)

+++ backend/src/main/java/com/dopamin/omok/game/physical/application/dto/PhysicalSnapshot.java
@@ -37,8 +37,8 @@ public record PhysicalSnapshot(
     public record PhysicalPlayerView(
             String color,             // BLACK / WHITE
             String nickname,
-            int x,
-            int y,
+            double x,                 // 칸 단위 위치(격자 모드는 정수값, 연속 모드는 소수값) — 클라가 보간해 렌더
+            double y,
             String heldItem,          // 보유 아이템 종류명(없으면 null)
             long destroyReadyAt,      // 파괴 쿨다운 해제 시각(epoch ms) — HUD 링 표시용
             boolean speedBoosted,
